requirements.txt
flask
gunicorn
pdfminer.six
spacy
python-dotenv
